% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_batch.R
\name{autoplot.batch}
\alias{autoplot.batch}
\title{Autoplot method for batch models}
\usage{
\method{autoplot}{batch}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'batch'}

\item{...}{further arguments are ignored}
}
\description{
Plots the most recent weights in each quantile using ggplot2.
}
