% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_online.R
\name{autoplot.online}
\alias{autoplot.online}
\title{Autoplot method for online models}
\usage{
\method{autoplot}{online}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'online'}

\item{...}{further arguments are ignored}
}
\description{
Plots the most recent weights in each quantile using ggplot2.
}
