% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conline_exports.R
\name{conline}
\alias{conline}
\title{Create an conline Object from the conline C++ Class}
\value{
A `conline` object from the _C++_ conline Class.
}
\description{
Allows for the creation of a Online Object in _C++_ from _R_
using the _C++_ conline class.
}
\examples{
conline_obj <- new(conline)
}
