% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{make_knots}
\alias{make_knots}
\title{Create a vector of knots for splines}
\usage{
make_knots(n, mu = 0.5, sig = 1, nonc = 0, tailw = 1, deg = 1)
}
\arguments{
\item{n}{Number of knots}

\item{mu}{Beta distribution location parameter}

\item{sig}{Beta distribution scale parameter}

\item{nonc}{Beta distribution noncentrality parameter}

\item{tailw}{Tailweight}

\item{deg}{Degree of splines}
}
\description{
This function creates a knot vector for splines. The knots are distributed
according to a beta distribution. The first input defines the number of inner
knots. The total number of knots is \code{n + 2 * order}.
}
