% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{oracle}
\alias{oracle}
\title{Probabilistic Forecast Combination - Oracle}
\usage{
oracle(y, experts, tau, intercept = FALSE, affine = FALSE, positive = FALSE, loss_function = "quantile",
loss_parameter = 1, forget = 0)
}
\arguments{
\item{y}{A numeric vector of realizations.}

\item{experts}{A an array of predictions with dimension
(Observations, Quantiles, Experts).}

\item{tau}{A numeric vector of probabilities.}

\item{intercept}{Determines if an intercept is added, defaults to FALSE. In this case a new expert is added as expert 1 allways predicting 1.}

\item{affine}{Defines wether weights are summing to 1 or now. Defaults to FALSE.}

\item{positive}{Defines wether weights have to positive. Defaults to FALSE.}

\item{loss_function}{Either "quantile", "expectile" or "percentage".}

\item{loss_parameter}{Optional parameter scaling the power of the loss function.}

\item{forget}{Adds an exponential forgetting to the optimization. Past observations will get less influence in the optimization. Defaults to 0 which corresponds to no forgetting.}
}
\value{
Returns weights and corresponding predictions.
It is possible to calculate the best convex combination
of weights by setting affine and positive to TRUE.
}
\description{
Returns predictions and weights calculated by numeric
optimization. The optimization is done in hindsight.
This means all observations are used.
}
