% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Pinball Loss}
\usage{
pinball_loss(real, pred, prob)
}
\arguments{
\item{real}{A numeric vector of realizations}

\item{pred}{A matrix of predictions. Columns
correspond to quantiles, rows to observations}

\item{prob}{A numeric vector of probabilities
corresponding to the columns of pred}
}
\description{
This function returns the pinball loss (also called quantile score)
of pointwise probabilistic predictions.
}
