% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_batch.R
\name{plot.batch}
\alias{plot.batch}
\title{Plot method for batch models}
\usage{
\method{plot}{batch}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from 'batch'}

\item{...}{further arguments are ignored}
}
\description{
Plots the most recent weights in each quantile.
}
