% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_plot.R
\name{plot.online}
\alias{plot.online}
\title{Plot method for online models}
\usage{
\method{plot}{online}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from 'online'}

\item{...}{further arguments are ignored}
}
\description{
Plots the most recent weights in each quantile.
}
