% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_online.R
\name{predict.online}
\alias{predict.online}
\title{Predict method for online models}
\usage{
\method{predict}{online}(object, new_experts, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'online'}

\item{new_experts}{new expert predictions}

\item{...}{further arguments are ignored}
}
\value{
\code{predict.online} produces an updated model object.
}
\description{
Calculates predictions based on new expert advice.
This does not update weights. If new observations are available
use update instead. The latter updates and computes predictions.
}
