% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_batch.R
\name{print.batch}
\alias{print.batch}
\title{Print method for batch models}
\usage{
\method{print}{batch}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from 'batch'}

\item{...}{further arguments are ignored}
}
\description{
Prints the average loss of all and the forecast combination.
}
