% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spline_fit}
\alias{spline_fit}
\title{Spline Fit - Fit Penalized B-Splines}
\usage{
spline_fit(y, x, lambda = 1, ndiff = 1L, deg = 3L, rel_nseg = 0.1)
}
\arguments{
\item{y}{The respones variable. Must be a numeric vector.}

\item{x}{Explanatory variable. Must be a numeric vector.}

\item{lambda}{The penalization parameter.}

\item{ndiff}{Degree of the difference operator to use when calculating
the B-Spline basis. A value of 1 corresponds to shrinkage towards
a constant, 2 corresponds to shrinkage towards a linear relationship
between x and y.}

\item{deg}{the degree of the basis functions.}

\item{rel_nseg}{The number of basis functions are calculated relative to the
length of y. A value of 0 corresponds to only 1 knot. A value of 1
corresponds to length(y) knots.}
}
\description{
Returns fitted values.
}
