% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_summary.R
\name{summary.online}
\alias{summary.online}
\title{Summary method for online models}
\usage{
\method{summary}{online}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'online'}

\item{...}{further arguments are ignored}
}
\description{
Calculates parameters chosen during optimization and aggregates losses.
}
