% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tidy.R
\name{tidy.online.experts_loss}
\alias{tidy.online.experts_loss}
\title{Tidy the Experts' losses of an Online object}
\usage{
\method{tidy}{online.experts_loss}(x, ...)
}
\arguments{
\item{x}{The experts_loss of an \code{online} object.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{t} \code{d} \code{p} \code{k} and \code{w} corresponding
to the time, marginals, probabilities, and experts_loss
of the online-learning computation.
}
\description{
\code{tidy} will transform the \code{experts_loss} array of an online object
into a tibble that is better suited for plotting and analysis.
}
