% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tidy.R
\name{tidy.online.predictions}
\alias{tidy.online.predictions}
\title{Tidy the Predictions of an Online object}
\usage{
\method{tidy}{online.predictions}(x, ...)
}
\arguments{
\item{x}{The predictions of an \code{online} object.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{t} \code{d} \code{p} \code{k} and \code{w} corresponding
to the time, marginals, probabilities, and predictions
of the online-learning computation.
}
\description{
\code{tidy} will transform the \code{predictions} array of an online object
into a tibble that is better suited for plotting and analysis.
}
