% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_tidy.R
\name{tidy.online.weights}
\alias{tidy.online.weights}
\title{Tidy the Weights of an Online object}
\usage{
\method{tidy}{online.weights}(x, ...)
}
\arguments{
\item{x}{The weights of an \code{online} object.}

\item{...}{Not currently used.}
}
\value{
A tibble with columns \code{t} \code{d} \code{p} \code{k} and \code{w} corresponding
to the time, marginals, probabilities, experts, and weights
of the online-learning computation.
}
\description{
\code{tidy} will transform the weights array of an online object
into a tibble that is better suited for plotting and analysis.
}
