% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_online.R
\name{update.online}
\alias{update.online}
\title{Update method for online models}
\usage{
\method{update}{online}(object, new_y, new_experts = as.numeric(c()), ...)
}
\arguments{
\item{object}{Object of class inheriting from 'online'}

\item{new_y}{new observations}

\item{new_experts}{new expert predictions. This must be left unspecified
if the model already contains the expert predictions corresponding to new_y.}

\item{...}{further arguments are ignored}
}
\value{
\code{update.online} produces an updated model object.
}
\description{
Continues learning using new observations and new expert advice.
}
