% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conline_exports.R
\name{conline}
\alias{conline}
\title{Create an conline Object from the conline C++ Class}
\value{
A \code{conline} object from the \emph{C++} conline Class.
}
\description{
Allows for the creation of a Online Object in \emph{C++} from \emph{R}
using the \emph{C++} conline class.
}
\examples{
conline_obj <- new(conline)
}
