% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_predict.R
\name{predict.online}
\alias{predict.online}
\title{Predict method for online models}
\usage{
\method{predict}{online}(object, new_experts, update_model = TRUE, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'online'}

\item{new_experts}{new expert predictions}

\item{update_model}{Defines wether the model object should be updated or not.
If TRUE, new forecaster and expert predictions are appended onto
the respective object items. Defaults to TRUE.}

\item{...}{further arguments are ignored}
}
\value{
\code{predict.online} produces an updated model object.
}
\description{
Calculates predictions based on new expert advice.
This does not update weights. If new observations are available
use update instead. The latter updates and weights and computes predictions.
}
