% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_print.R
\name{print.online}
\alias{print.online}
\title{Print method for online models}
\usage{
\method{print}{online}(x, ...)
}
\arguments{
\item{x}{Object of class inheriting from 'online'}

\item{...}{further arguments are ignored}
}
\description{
Prints the average loss of all experts and the forecast combination.
}
