\name{profoc-package}
\alias{profoc-package}
\docType{package}
\title{
Package Info
}
\description{
Use multiple online-aggregation algorithms to combine probabilistic forecasts using CRPS Learning as described in Berrisch, Ziel: "CRPS Learning", 2021. The primary function of this package is called profoc.
}
\details{

\packageIndices{profoc}

}
\author{

Maintainer: Jonathan Berrisch \url{mailto:Jonathan@Berrisch.biz}

Co-Author: Florian Ziel

}
\references{
Berrisch, Ziel: "CRPS Learning", 2021
}

\keyword{ package }
\seealso{

Source Code: \url{https://github.com/BerriJ/profoc}

BugReports: \url{https://github.com/BerriJ/profoc/issues}

}